Move user tests into user package
